requirements.txt
streamlit==1.32.0
google-generativeai
sqlalchemy
psycopg2-binary
PyPDF2
pandas